Modify the final state criterion for Asm
Instead of simply testing whether SP is equal to its initial value,
we use the more general criterion that SP be a valid pointer of the
initial memory.

This is necessary for compositionality, because it remains accurate
in the case of nested cross-component tail calls (for example),
where the inner component returns directly from the top-level call.