Genv.globalenv: operate in terms of prog_defmap
The match_prog relation used by Unusedglobproof operates in terms of
prog_defmap so using the same starting point for Genv.globalenv will
make it easier to generalize the common match_program_gen relation
to cover Unusedglobproof.

This also cleans up the interface and we use it to prove the relation
between linkorder and Genv.globalenv used by AsmLinking.