New approach to Stackingproof
There are two aspects of the Stacking pass which conspire
to make things difficult. One one hand, the stacking simulation
relation is fairly complex to begin with. On the other hand,
for various reasons splitting the simulation convention into
a structural component and a CKLR component is more difficult
than with other translation passes.

To make this easier to deal with, I separate concerns by
verifying Stacking against a monolithic simulation convention
[cc_stacking], then prove a correspondance between that
simulation convention and the ones we use for commutation
properties.